[MOB-16085] -- Move common freemarker's logic into separate module
