+++ core/hh-freemarker-wrapper/build.gradle.kts
@@ -2,6 +2,7 @@ plugins {
     id(GradlePlugins.gradleIntelliJPlugin)
     kotlin("jvm")
     id(GradlePlugins.setupIdeaPlugin)
+    id(GradlePlugins.coreModuleMarker)
 }
 
 repositories {
@@ -9,10 +10,6 @@ repositories {
 }
 
 dependencies {
-    implementation(project(":hh-plugins-core"))
-    implementation(project(":hh-freemarker-wrapper"))
-
     implementation(kotlin("stdlib-jdk8"))
-    implementation(kotlin("reflect"))
     implementation(Libs.freemarker)
 }
