+++ core/hh-freemarker-wrapper/src/main/kotlin/ru/hh/plugins/freemarker_wrapper/adapters/IntPropertyAdapter.kt
@@ -1,4 +1,4 @@
-package ru.hh.plugins.utils.freemarker.adapters
+package ru.hh.plugins.freemarker_wrapper.adapters
 
 import com.android.tools.idea.observable.core.IntValueProperty
 import freemarker.template.*
