+++ plugins/hh-carnival/build.gradle.kts
@@ -11,6 +11,8 @@ repositories {
 
 dependencies {
     implementation(project(":hh-plugins-core"))
+    implementation(project(":hh-freemarker-wrapper"))
+
     implementation(kotlin("stdlib-jdk8"))
     implementation(kotlin("reflect"))
     implementation(Libs.freemarker)
